Add 4 rotary encoder modes by inverting the A/B signals
Detect when the A/B signals have been inverted by tracking when the encoder
is active for a long time, which implies that the idle level has changed.

Require a full sequence of encoder signal events before detecting an
adjustment otherwise A/B signal inversions can form part of a rotation
event.

This doesn't work because it's easy to linger with the encoder partially
active when changing direction which will change the rotary mode and then
there's no way to identify that the movements are offset because all of the
timings are variable.

Increasing the delay required before a mode change takes effect will not
solve the problem because it'll take too long to change modes. When the
wrong mode is detected it'll then require an even longer wait to recover.